# [Logs] Use symbolic links to stdout instead of files