--- a/bin/compile
+++ b/bin/compile
@@ -277,7 +277,7 @@ if [ "$(php_api_version)" = "20170718" -o "$(php_api_version)" = "20180731" ] ; 
     fetch_package "ext/$(php_api_version)/php-mcrypt" "/app/vendor/php" | indent
 fi
 
-if [ -f "$BUILD_DIR/composer.json" ] && package_newrelic_enabled; then
+if [ -f "$BUILD_DIR/composer.json" ] && package_logsnewrelic_enabled; then
     LOG_FILES+=( "/app/vendor/newrelic/daemon.log" "/app/vendor/newrelic/agent.log" )
     install_newrelic "$NEWRELIC_VERSION"
 fi
@@ -457,8 +457,7 @@ else
     echo "Using WEB_CONCURRENCY=\${WEB_CONCURRENCY} processes." >&2
 fi
 
-`init_log_plex ${LOG_FILES}`
-`tail_log_plex ${LOG_FILES} ${SYS_LOG_FILES}`
+`init_stdout ${LOG_FILES}`
 
 (
     php-fpm -p "/app/vendor/php"