--- a/bin/compile
+++ b/bin/compile
@@ -277,7 +277,7 @@ if [ "$(php_api_version)" = "20170718" -o "$(php_api_version)" = "20180731" ] ; 
     fetch_package "ext/$(php_api_version)/php-mcrypt" "/app/vendor/php" | indent
 fi
 
-if [ -f "$BUILD_DIR/composer.json" ] && package_logsnewrelic_enabled; then
+if [ -f "$BUILD_DIR/composer.json" ] && package_newrelic_enabled; then
     LOG_FILES+=( "/app/vendor/newrelic/daemon.log" "/app/vendor/newrelic/agent.log" )
     install_newrelic "$NEWRELIC_VERSION"
 fi