--- a/support/package_reader
+++ b/support/package_reader
@@ -1,0 +1,15 @@
+#!/usr/bin/env bash
+
+basedir="$( cd -P "$( dirname "$0" )" && pwd )"
+
+source "$basedir/../conf/buildpack.conf"
+source "$basedir/lib/utils"
+source "$basedir/lib/swift"
+
+
+pushd $basedir/reader
+go build .
+popd
+
+swift_upload reader/reader
+rm reader/reader