@johnsudaar 2020-09-10
Impact on the perfs:
New system:
```
time bash -c 'for i in $(seq 1000); do curl "https://MY_APP.osc-fr1.scalingo.io/" --silent > /dev/null && echo -n "." || echo -n "X" ; done'
........................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................bash -c   8.45s user 3.33s system 7% cpu 2:37.43 total
```

Old system:

```
time bash -c 'for i in $(seq 1000); do curl "https://test-php.osc-fr1.scalingo.io/" --silent > /dev/null && echo -n "." || echo -n "X" ; done'
.............................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
...........................................bash -c   7.99s user 3.16s system 6% cpu 2:55.01 total
```
So the new solution is even a bit faster than the old one \o/